Add files via upload
This is a demonstration of the use of MQTT protocol with a cloud computer.
Raspberry Pi with node-red running the graphics and the MQTT protocol that is subscribed to the topic name temperature on the broker, which is a virtual machine on Google Cloud.
The temperature sensor is connected to an ESP that updates the value to the topic name temperature on the same virtual machine on Google Cloud. 
As these devices are connected through the cloud computer.
They have the freedom to connect through any individual internet connection to update and read the temperature value from anywhere in the world. 
Using the MQTT protocol makes the devices connected in a simple way. This is the code of the ESP truck side, and in this whole code, there is no requirement to know who is reading the data.
The only purpose is to update the sensor value on the topic name with the IP address of the virtual machine acting as a broker. 

Devices that subscribe to the topic temperature use the same virtual machine IP address to read the data without knowing who is updating it. 
and this is the virtual machine google server account that shows the ip address used by both esp and resberypai to access topic name temperature.
And multiple devices can read and update the value of the same topic.